+++ Actions/Tools/SpindleTool/NoneSpindleToolFactory.cs
@@ -7,7 +7,7 @@ namespace Actions.Tools.SpindleTool
 {
     public class NoneSpindleToolFactory : ISpindleToolFactory
     {
-        public ModbusToolCommand CreateSpindleToolCommand(SpindleState.SpindleRotationState rotation, double speed)
+        public ModbusToolCommand CreateSpindleToolCommand(SpindleState.SpindleRotationState rotation, decimal speed)
         {
             return new ModbusToolCommand { Registers = new ModbusRegister[0], Delay = 0 };
         }
